requirements.txt
sqlalchemy >= 1.0.0
mysql-connector >= 2.1.3

flask >= 0.11
flask_cors >= 2.1.2
Flask-RESTful >= 0.3.5

#Openstack Kilo release
python-novaclient >= 2.23.2
python-troveclient >= 1.0.9

paramiko >= 2.0.2
vertica_python >= 0.6.5
schedule >= 0.3.2
futures >= 3.0.5
